Migrated CAT to a separate crate.
