+++ crates/tool-cat/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cat"
+version = "1.0.0"
+edition = "2021"
+authors = ["Breno RdV"]
+description = "Port of the Unix cat command."
+repository = "https://github.com/brenordv/rusted-toolbox"
+
+[dependencies]
+shared = { path = "../shared" }
+clap = "4.5.45"
+anyhow = "1.0.99"
+rstest = "0.26.1"
+tempfile = "3.20.0"
+tracing = "0.1.41"
