+++ crates/tool-cat/src/cat_app.rs
@@ -1,9 +1,9 @@
-use crate::tools::cat::models::CatOptions;
 use anyhow::{Context, Result};
 use shared::constants::general::SIZE_128KB;
 use std::fs::File;
 use std::io;
 use std::io::{BufRead, BufReader, ErrorKind, Read, Write};
+use crate::models::CatOptions;
 
 /// Processes file content with optional formatting.
 ///
@@ -167,7 +167,6 @@ fn raw_cat(path: Option<&str>) -> Result<()> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::cat::models::CatOptions;
     use rstest::*;
     use std::io::Write;
     use tempfile::NamedTempFile;
