--- crates/tools/src/tools/cat/mod.rs
@@ -1,3 +1,0 @@
-pub mod cat_app;
-pub mod cli_utils;
-pub mod models;
